+++ src/components/MapView/Direction/index.native.tsx
@@ -1,6 +1,6 @@
 import Mapbox from '@rnmapbox/maps';
-import {DirectionProps} from './MapViewTypes';
-import styles from '../../styles/styles';
+import {DirectionProps} from '../MapViewTypes';
+import styles from '../../../styles/styles';
 
 function Direction({coordinates}: DirectionProps) {
     if (coordinates.length < 1) {
